move route indicators component to it's own file
